fix(ios): split PlatformFeatures to satisfy file_length lint
LMSFeatureModels.swift exceeded SwiftLint's 1000-line error threshold
after M9 feature flags were added. Move PlatformFeatures into its own
file following the existing LMSFeatureModels* split pattern.
